Moved peer listener and connect from server to client side
