+++ app/src/main/java/com/psu/capstonew17/backend/sharing/SharingManager.java
@@ -1,19 +1,23 @@
 package com.psu.capstonew17.backend.sharing;
-
+import android.content.BroadcastReceiver;
+import android.content.Context;
 import android.graphics.Bitmap;
-
+import android.net.wifi.WpsInfo;
+import android.net.wifi.p2p.WifiP2pConfig;
+import android.net.wifi.p2p.WifiP2pDevice;
+import android.net.wifi.p2p.WifiP2pDeviceList;
+import android.net.wifi.p2p.WifiP2pManager;
 import com.google.zxing.WriterException;
 import com.google.zxing.common.BitMatrix;
 import com.google.zxing.qrcode.QRCodeWriter;
 import com.google.zxing.qrcode.encoder.Encoder;
 import com.google.zxing.qrcode.encoder.QRCode;
 import com.google.zxing.BarcodeFormat;
-
 import com.psu.capstonew17.backend.api.Card;
 import com.psu.capstonew17.backend.api.Deck;
 import com.psu.capstonew17.backend.api.SharingReceiveListener;
 import com.psu.capstonew17.backend.api.SharingTransmitListener;
-
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 
@@ -22,6 +26,20 @@ import java.util.Random;
  */
 public class SharingManager implements com.psu.capstonew17.backend.api.SharingManager {
     private static SharingManager instance = null;
+    private List <WifiP2pDevice> listOfPeers = new ArrayList<WifiP2pDevice>();         //host list of peers from PeerListListener
+    private WifiP2pManager wifiManager;
+    private WifiP2pManager.Channel wifiChannel;
+    private BroadcastReceiver wifiReceiver;
+
+    //find peers to connect with
+    private WifiP2pManager.PeerListListener peerListener = new WifiP2pManager.PeerListListener() {
+        @Override
+        public void onPeersAvailable(WifiP2pDeviceList wifiP2pDeviceList) {
+            listOfPeers.clear();    //empty peer list
+            listOfPeers.addAll(wifiP2pDeviceList.getDeviceList());  //repopulate with new devices
+            if(listOfPeers.size() ==0){System.out.println("No devices within range");}  //need to modify for android device
+        }
+    };
 
     public static class LinkParameters {
         String ssid;
@@ -93,8 +111,27 @@ public class SharingManager implements com.psu.capstonew17.backend.api.SharingMa
         return bmap;
     }
 
+    public void boradcastReceive(){
+        //TODO: implement broadcast reciever for client side
+    }
 
     @Override
     public void receive(String code, RxOptions opts, SharingReceiveListener listener) {
+        WifiP2pDevice aDevice = listOfPeers.get(0); //get device in front of list
+        WifiP2pConfig configuration = new WifiP2pConfig();
+        configuration.deviceAddress = aDevice.deviceAddress;    //MAC address IDing device
+        configuration.wps.setup= WpsInfo.PBC;           //wifi protected setup push button config
+        wifiManager.connect(wifiChannel, configuration, new WifiP2pManager.ActionListener() {
+            //will using our own callback routine insted of WifiP2pManager.ActionListener()
+            @Override
+            public void onSuccess() {
+                //if/else cases in broadcast reciver will provide noitificaiton for successful connect
+            }
+
+            @Override
+            public void onFailure(int i) {
+                System.out.println("Failed to connect");
+            }
+        });
     }
 }
